Frontend-Anbindung ans Backend: Fehler gemacht, Testläufe zur Ursachenanalyse
Fehlersuche: falsche Pfade im Frontend korrigiert, weitere Tests

Mixed-Content/CORS getestet, Pfade angepasst

+++ src/main/java/br/com/ruana/mediaflix/controller/SerieController.java
@@ -46,7 +46,7 @@ public class SerieController {
     }
 
     @GetMapping("/{id}/temporadas/{numero}")
-    public List<EpisodioDTO> obterTemporadasPorNumero(@PathVariable Long id, @PathVariable Long numero){
+    public List<EpisodioDTO> obterTemporadasPorNumero(@PathVariable Long id, @PathVariable Integer numero){
         return servico.obterTemporadasPorNumero(id, numero);
     }
