+++ src/main/java/br/com/ruana/mediaflix/repository/SerieRepository.java
@@ -39,5 +39,5 @@ public interface SerieRepository extends JpaRepository<Serie, Long> {
     List<Serie> lancamentosMaisRecentes(PageRequest pageable);
 
     @Query("SELECT e FROM Serie s JOIN s.episodios e WHERE s.id = :id AND e.temporada = :numero")
-    List<Episodio> obterEpisodiosPorTemporada(Long id, Long numero);
+    List<Episodio> obterEpisodiosPorTemporada(Long id, Integer numero);
 }
